separate out ReturnJourney
