+++ app/sdil/controllers/SdilWMController.scala
@@ -332,24 +332,6 @@ trait SdilWMController extends WebMonadController
 
   implicit val address: Format[SmallProducer] = Json.format[SmallProducer]
 
-  def askReturn(subscription: RetrievedSubscription, sdilRef: String, sdilConnector: SoftDrinksIndustryLevyConnector)(implicit hc: HeaderCarrier): WebMonad[SdilReturn] = for {
-    ownBrands      <- askEmptyOption(litreagePair, "own-brands-packaged-at-own-sites") emptyUnless !subscription.activity.smallProducer
-    contractPacked <- askEmptyOption(litreagePair, "packaged-as-a-contract-packer")
-    askSmallProd   <- ask(bool, "exemptions-for-small-producers")
-    firstSmallProd <- ask(smallProducer(sdilRef, sdilConnector), "first-small-producer-details") when askSmallProd
-    smallProds     <- manyT("small-producer-details",
-      {ask(smallProducer(sdilRef, sdilConnector), _)},
-      min = 1,
-      default = firstSmallProd.fold(List.empty[SmallProducer])(x => List(x)),
-      editSingleForm = Some((smallProducer(sdilRef, sdilConnector), smallProducerForm))
-    ) when askSmallProd
-    imports        <- askEmptyOption(litreagePair, "brought-into-uk")
-    importsSmall   <- askEmptyOption(litreagePair, "brought-into-uk-from-small-producers")
-    exportCredits  <- askEmptyOption(litreagePair, "claim-credits-for-exports")
-    wastage        <- askEmptyOption(litreagePair, "claim-credits-for-lost-damaged")
-    sdilReturn     =  SdilReturn(ownBrands,contractPacked,smallProds.getOrElse(Nil),imports,importsSmall,exportCredits,wastage)
-  } yield sdilReturn
-
   implicit val smallProducerHtml: HtmlShow[SmallProducer] =
     HtmlShow.instance { producer =>
       Html(producer.alias.map { x =>
